+++ src/core/solFileResolvers.ts
@@ -2,6 +2,7 @@ import fs from "fs"
 import { ContractPackage } from "@eversdk/appkit"
 import { findExisting } from "./utils"
 import { SOLIDITY_FILE } from "../controllers/solidity"
+import os from "os"
 
 type ResolvedContractPackage = {
     package: ContractPackage
@@ -10,6 +11,9 @@ type ResolvedContractPackage = {
 }
 
 export function resolveContract(filePath: string): ResolvedContractPackage {
+    if (filePath.startsWith("~")) {
+        filePath = `${os.homedir()}${filePath.substring(1)}`
+    }
     filePath = filePath.trim()
     const lowered = filePath.toLowerCase()
     let basePath
@@ -35,7 +39,9 @@ export function resolveContract(filePath: string): ResolvedContractPackage {
             ? JSON.parse(fs.readFileSync(abiPath, "utf8"))
             : undefined
     if (!abi) {
-        throw new Error("ABI file missing.")
+        throw new Error(
+            `You have specified "${filePath}" as a contract file name, but a corresponding ABI file is missing. ABI file must have an extension ".abi" or ".abi.json".`,
+        )
     }
     return {
         package: {